adapter: Early p2p binding
The LLCP wks array will be properly built before the MAC link is up by
binding to p2p services as soon as a p2p driver is registered.
The bound will also happen when the MAC link is up for supporting kernels
where the bound LLCP sockets will be closed when the MAC is brought down.
For more recent kernels, the bind operation will happen only once.