test: Add handover_agent for testing
This is a very simple implementation of handover agent useful for
testing (especially if not willing to run BlueZ 5.x). Currently
received data in PushOOB/RequestOOB is printed and reply to RequestOOB
is generated from static data.

In future this might be extended to accept some user input and more
flexible reply generation.