doc: Add 'State' field to handover agent API
This will allow both agent and neard to be informed about power state
of carrier. It is up to agent to decide what to do with information
provided e.g. in case of 'activating' state it could delay connection
attempt. This will also allow to properly set Carrier Power State in
ac record.