p2p: Support for single connection server
The LLCP validation tests require to reject any connection once a client
is connected. Which means a new listen should start iff the current
connection is closed. For that a single_connection field is added to the
p2p_driver structure and the p2p_bind code is factorized so that the
client close hook can call it without re-allocating a server data pointer.