unit: Add test for Handover Select
This adds test for HS. Currently test is not enabled as it is not
possible to run HO code as unit test now as this code does both
parsing and acting (eg. trying to access HO agent) in same run.

This test is added for reference and should be enabled when handover
code is splitted into parsing and acting functions.