iso15693: Add plugin for ISO/IEC 15693 Type 5 Tags
ISO/IEC 15693 tags that are properly NDEF formatted are referred to
as Type 5 tags (they are no longer referred to as Type V tags). The
NDEF Capabilities Container (CC) for Type 5 tags starts on the first
byte of the first block of user memory.  The CC is 4 bytes long but
the NFC-V committee is considering adding an 8-byte CC when the amount
of user memory is larger than what can be represented in the 4-byte CC.

The most recent Type 5 tag draft specification states that the NDEF
data area will start on the first byte following the last CC byte. The
NFC-V committee is considering changing that to the beginning of the
next block after the last CC byte.  The code being added implements
that option--most tags seem to be have a 4-byte block size anyway so
its not an issue.

Before adding the actual read, write, check_presence, and format
functionality, add some helper routines that will be used by the code
that implements that functionality.
Also add the hooks so that the Type 5 tag code get built.