iso15693: Add support for Type 5 tag check_presence
Add support for the check_presence hook for Type 5 tags. To check
whether a tag is still present, simply read the very first block of
the tag and if it succeeds, the tag must still be present. Otherwise,
something serious is wrong or that tag is no longer present. Either
way its bad so make the check_presence() call report an error.