iso15693: Add support for Type 5 tag formatting
Add support for formatting Type 5 tags.  To accomplish this, the
Type 5 NDEF Capabilities Container (CC) will be written to the very
beginning of the tag's user memory.  The CC is referred to as the "META"
area in the code (as is done in the code for other tag types).

One of the flags in the CC is whether the tag supports the ISO/IEC 15693
'Read Multiple Blocks' command.  To determine this, the code issues the
command and if it fails, clear the flag in the CC; if it succeeds, then
it's obviously supported.

The flag indicating whether 'Lock' commands are supported is not set
since the only way to tell if its supported is to issue that command.
The problem is that the lock is permanent so we can't do that.

Also, the flag indicating whether special framing is required is not set
either since I have no information as to what that means and, if special
framing is required, the write of the CC will probably fail anyway.

It should be noted that byte 3 of the CC defined in the current draft of
the Type 5 tag specification is slightly different from the previous
definition for Type V tags that I've seen. The differences are shown below.

		Previous		Type 5
		--------		------
Bits 7-5:	 RFU			 RFU
Bit 4:	 RFU			 Special Frame
Bit 3:	 RFU			 Lock supported
Bit 2:	 CC2 Overflow		 RFU
Bit 1:	 RFU			 RFU
Bit 0:	 MBRead Supported	 MBRead Supported