nfctype4: Only do entire NDEF detection procedure once
The current Type 4 read code goes through the
entire NDEF detection procedure regardless of
whether its been gone through before or not.

This causes a problem when the tag is re-read
immediately after a write operation because
the tag is already selected.  What happens is
the tag responds to the select attempts that
are a part of the second NDEF detection procedure
with a 0x6a82 (NDEF Tag Application not found)
response.  This leaves things in an odd state
where the tag has been written to but neard and
its clients can't see the new data because the
subsequent read operation failed.

Fix this by making nfctype4_read() check whether
the tag is already in the selected state before
starting the NDEF detection procedure.  If it is
selected, we can skip part of the NDEF detection
procedure and go directly to reading the NDEF file
size.

Use the value returned by near_tag_get_r_apdu_max_size()
to indicate whether the tag is in the selected state.
If the returned value is 0, its not in the selected
state.  Anything other than 0 means that the tag is
currently selected.

To ensure that the value returned by
near_tag_get_r_apdu_max_size() is an accurate
reflection of the selected state, we have to set the
value to 0 when the select fails in select_ndef_file().
For completeness, also set the values returned by
near_tag_set_max_ndef_size() and
near_tag_set_c_apdu_max_size() to 0 as well.