nfctype4: Handle NDEF file larger than Max R-APDU Data Size"
The current t4_readbin_NDEF_ID() implementation assumes that the NDEF
file length is less than or equal to the Max R-APDU Data Size. This
may not alway be true so be sure to read the lesser of the two values.
data_read_cb() will continue to read the rest of the data when the NDEF
file length is larger than the Max R-APDU Data Size.