nfctype4: Use correct P2 value for 'Select' commands
The current implementation of the ISO_Select() routine always sets the
P2 value for the 'Select' command to 0. However, according the the NFC
Forum, Type 4 Tag Operation Specification, the value of P2 depends on
what is being selected.  The proper values from the specification are
shown below.

  Target of Selection	 P1	 P2	Spec Reference
  -------------------	 --	 --	--------------
  NDEF Tag Application	0x04	0x00	Table 10
  Capability Container	0x00	0x0C	Table 13
  NDEF File			0x00	0x0C	Table 19

So, when P1 is 0x04, P2 is 0x00 and when P1 is 0x00, P2 is 0x0C.  Add
this logic to the ISO_Select() routine so the proper value for P2 is
always used.