nfctype4: check_presence should select NDEF file
The nfctype4_check_presence() routine should select the NDEF file
instead of the NDEF Tag Application otherwise write operations may
fail. To accomplish this, store the the NDEF file ID in the tag
structure and use it when issuing the 'Select' command thats a
part of the check presence process.