se: Use callback function to handle the request data
Due to we use queue to save the command, the callback function will handle
the error operation, else it may lead to buffer double free.

Move the free operation of req->apdu to io_cb since the io_cb function
may be called directly in the plugin, after req struct is freed,
req->apdu will be a wild pointer.