ace: Add length option for get refresh tag operation
In command INS_GET_GP_DATA, when try to get the refresh tag
value, the return value is not null, then we need to set the
resp_lenght to 0. Also the response value for get refresh tag
operation is (tag, length, value), we missed the length option
in current code.