adapter: Prioritize p2p when discovering multiple targets
In NCI it is possible for the NFCC to discover multiple targets.
In that case the kernel part will send a netlink event for each
of them before sending an end of list one.

In the current implementation the connection to the target is done
before the target done event. So the first discovered target was
the one that neard tries to connect to.

This patch fixes that by enabling p2p targets before tags or SEs ones.
We probably want to connect to NFC-DEP (usualy having NFC-DEP and
ISO-DEP on the same device means that ISO-DEP is the secure element
interface).