nfctype1: Fix dynamic tag read
The current implementation does not handle type1 dynamic tag reads
correctly, the dynamic command size format is different as the data
field is 8 bytes long (As opposed to 1 byte for the static commands).
Also, lock and reserved TLVs are not parsed so when several commands
are needed to retrieve all chunks of memory, lock and reserved
area are also copied, building a corrupted NDEF.
This is fixed by skipping all those areas when copying the NDEF over.