build: Use LT_INIT instead of AC_DISABLE_STATIC and AC_PROG_LIBTOOL
AC_PROG_LIBTOOL is deprecated and wrongly uses the autoconf namespace,
use LT_INIT instead. As we are using LT_INIT anyway we might as well
pass the disable-static option directly to it instead of using the
AC_DISABLE_STATIC macro, which is not deprecated right now, but uses
the autoconf namespace too.