build: Unify and centralize plugindir default to be PKGLIBDIR/plugins
Unify its default value to be under PKGLIBDIR, instead of LIBDIR/near,
which does not match the name of the project nor the daemon. This also
fixes a discrepancy between the directory encoded in the daemon and
the pkg-config file. Centralize the variable to be only set once in
configure.ac.