nfctool: polling: Fix netlink event handler
With this patch, the netlink event handler for poll operations now only
treats messages from the device specified with the -d command line
parameter.

This fixes an issue when used with nfcsim (the loopback driver) for
which you have 2 instances of nfctool running simultaneously in poll
mode. Both instances were receiving the TARGETS_FOUND and TM_ACTIVATED
events which could lead to prematurely stop the poll operation for the
target device.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>