tag: Add Tag deactivate support
Once a tag has been discovered, it remains active until it
is moved out of range, or an error occurs while reading or
writing it.  While the tag is active, the adapter cannot
be powered off.  This wastes power when the client application
no longer cares whether the tag is in range or not.

To address this issue, add the ability for a client application
to deactivate an active tag.  Once deactivated, the client
application can poll the tag again to read it or power off the
adapter.

The 'NFC_CMD_DEACTIVATE_TARGET' netlink command is added to
request that the kernel deactivate the target (tag).

Signed-off-by: Mark Greer <mgreer@animalcreek.com>