nfctype5: Add support for TI Standard and Pro Type 5 tag
Standard and Pro Type 5 tags from Texas Instruments do not support
the Get System Information command which means they cannot be read
with the current neard code.  Fortunately, both types of tags have
eight, 4-byte blocks so that information can be filled in instead
of issuing the Get System Information command to get it.  With this
change, Standard and Pro tags can now be formatted, read, and written.

Signed-off-by: Mark Greer <mgreer@animalcreek.com>